requirements.txt
Flask>=3.0
Pillow>=10.0
reportlab>=4.0
openpyxl>=3.1
python-dotenv>=1.0
